project/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
web3==6.11.3
tensorflow==2.15.0
supabase==2.4.3
python-dotenv==1.0.0
PyJWT==2.8.1
oracledb==2.2.0
werkzeug==3.0.0
